CSV parsers refactoring
* Moved ParserInterface higher up in the hierarchy
* Created AbstractParser class
* Created AbstractCsvParser clas
* Updated CSV parsers to extend AbstractCsvParser class
* Moved common CSV code into the AbstractCsvParser class
